Change DataRepository/Collection nomenclature.
This is a tricky change, but an important one - prior to this change,
"DataRepository" was sometimes used to mean a concept that now maps
to our "Collection" concept, but sometimes seemed to refer more to a
Datastore storage backend, and sometimes seemed to refer to what we now
mean by "DataRepository", which is a thing that aggregates multiple
collections and (in general) multiple storage systems.

That said, most of the changes here are pretty obvious (to me, having
been part of the team that wrote them, at least), and they should be
consistent with the descriptions and the use cases these requirements
are derived from on LDM-592.